requirements.txt
Django>=4.2.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
django-extensions>=3.2.0
werkzeug>=2.3.0
pyOpenSSL>=23.0.0
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0

